# src/main/java/kz/halykacademy/bookstore/dto/AuthorDTO.java
package kz.halykacademy.bookstore.dto;

import java.util.List;

public class AuthorDTO {
    private Long id;
    private String name;
    private String surname;
    private List<BookDTO> books;
    private List<GenreDTO> genres;

    public AuthorDTO() {
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSurname() {
        return surname;
    }

    public void setSurname(String surname) {
        this.surname = surname;
    }

    public List<BookDTO> getBooks() {
        return books;
    }

    public void setBooks(List<BookDTO> books) {
        this.books = books;
    }

    public List<GenreDTO> getGenres() {
        return genres;
    }

    public void setGenres(List<GenreDTO> genres) {
        this.genres = genres;
    }
}